Tanweb: - Security rework to singleton - adjusted Scheduler to change
Blockers:
a) classes allowing to block operations for specific conditions,
	example: users cannot add schedule entries for previous month
	after day set in config.ini file (same for inspector)
a) Blockers
- "enum" class, for selecting config value for blocker
- for now 2 values, lockSchedule and lockInspector
b) OperationBlocker
- abstract for blockers
- passes main method "isBlocked"
- value of $input is passed to implemented classes
- method "shouldBeBlocked" determines if conditions to block are met
c) ScheduleBlocker
- class to block adding Schedule entries
- only admin and schedule_admin are always allowed
d) InspectorDateBlocker
- class to block Inspector module operations
- it blocks MOST of operations for inspector

Schedule:
- replaced old blocking code with ScheduleBlocker

Inspector:
- added blocker for Art 41
- added blocker for Decisions
- added blocker for Tickets
- added blocker for Suspensions
- added blocker for Suspension Snactions (! it uses sanction date !)

Fixes:
- Suspensions table correcly shows correction shift and headers in inspector module

+++ sys/classes/Controllers/InspectorArticle.php
@@ -11,6 +11,7 @@ use Tanweb\Container as Container;
 use Services\ArticleService as ArticleService;
 use Services\DocumentService as DocumentService;
 use Tanweb\Config\INI\Languages as Languages;
+use Custom\Blockers\InspectorDateBlocker as InspectorDateBlocker;
 
 /**
  * Description of InspectorArticle
@@ -52,29 +53,47 @@ class InspectorArticle extends Controller{
     
     public function saveNewArticle(){
         $data = $this->getRequestData();
-        $id = $this->articleService->saveNewArticle($data);
+        $languages = Languages::getInstance();
+        $blocker = new InspectorDateBlocker();
+        if($blocker->isBLocked($data)){
+            $this->throwException($languages->get('cannot_change_selected_month'));
+        }
+        else{
+            $id = $this->articleService->saveNewArticle($data);
+        }
         $response = new Container();
         $response->add($id, 'id');
-        $languages = Languages::getInstance();
         $response->add($languages->get('changes_saved'), 'message');
         $this->setResponse($response);
     }
     
     public function updateArticle(){
         $data = $this->getRequestData();
-        $this->articleService->updateArticle($data);
-        $response = new Container();
         $languages = Languages::getInstance();
+        $blocker = new InspectorDateBlocker();
+        if($blocker->isBLocked($data)){
+            $this->throwException($languages->get('cannot_change_selected_month'));
+        }
+        else{
+            $this->articleService->updateArticle($data);
+        }
+        $response = new Container();
         $response->add($languages->get('changes_saved'), 'message');
         $this->setResponse($response);
     }
     
     public function removeArticle(){
         $data = $this->getRequestData();
-        $id = (int) $data->get('id');
-        $this->articleService->removeArticle($id);
-        $response = new Container();
         $languages = Languages::getInstance();
+        $id = (int) $data->get('id');
+        $blocker = new InspectorDateBlocker();
+        if($blocker->isBLocked($data)){
+            $this->throwException($languages->get('cannot_change_selected_month'));
+        }
+        else{
+            $this->articleService->removeArticle($id);
+        }
+        $response = new Container();
         $response->add($languages->get('changes_saved'), 'message');
         $this->setResponse($response);
     }
