+++ sys/classes/Controllers/InspectorDecisions.php
@@ -12,6 +12,7 @@ use Services\DocumentService as DocumentService;
 use Services\SuspensionService as SuspensionService;
 use Tanweb\Container as Container;
 use Tanweb\Config\INI\Languages as Languages;
+use Custom\Blockers\InspectorDateBlocker as InspectorDateBlocker;
 
 /**
  * Description of InspectorDecisions
@@ -64,10 +65,16 @@ class InspectorDecisions extends Controller{
     
     public function saveDecision() : void {
         $data = $this->getRequestData();
-        $id = $this->decisionService->saveDecisionForCurrentUser($data);
+        $languages = Languages::getInstance();
+        $blocker = new InspectorDateBlocker();
+        if($blocker->isBLocked($data)){
+            $this->throwException($languages->get('cannot_change_selected_month'));
+        }
+        else{
+            $id = $this->decisionService->saveDecisionForCurrentUser($data);
+        }
         $response = new Container();
         $response->add($id, 'id');
-        $languages = Languages::getInstance();
         $response->add($languages->get('changes_saved'), 'message');
         $this->setResponse($response);
     }
@@ -75,9 +82,15 @@ class InspectorDecisions extends Controller{
     public function removeDecision() : void {
         $data = $this->getRequestData();
         $id = (int) $data->get('id');
-        $this->decisionService->removeDecision($id);
-        $response = new Container();
         $languages = Languages::getInstance();
+        $blocker = new InspectorDateBlocker();
+        if($blocker->isBLocked($data)){
+            $this->throwException($languages->get('cannot_change_selected_month'));
+        }
+        else{
+            $this->decisionService->removeDecision($id);
+        }
+        $response = new Container();
         $response->add($languages->get('changes_saved'), 'message');
         $this->setResponse($response);
     }
